Feature #30: Types, shield, action sequence, more tags in uppaal
We've implemented more of our own types. These are classes holding the state(containing the map, drone index in the map and various other variables), learning parameters(containing costs for turning, cost for moving and reward for discovery of new territory) and lastly drone specs.

Also implemented shield function that works the same as in uppaal, but on the newest map, thus trying to ensure safety (not working all the time).

STOMPC now uses an action sequence allowing the drone to make several actions before having to learn a new sequence.

With the addition of the new types in STOPMC, the uppaal model have been updated to use the information from objects containing information uppaal needs. This allows us to make changes quickly in python, without having to change anything in uppaal.

Co-authored-by: Thomas Lauritsen <Laurits1994@gmail.com>